管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -31,7 +31,8 @@ public class EmployeeAction extends ActionBase {
 
 //一覧画面に表示
 public void index() throws ServletException, IOException{
-
+  //管理者かどうかのチェック //追記
+    if (checkAdmin()) { //追記
 int page = getPage();
 List<EmployeeView> employees = service.getPerPage(page);
 
@@ -52,19 +53,22 @@ if (flush != null) {
 
 forward(ForwardConst.FW_EMP_INDEX);
 }
-
+}
 
 //新規登録画面を表示する
 public void entryNew() throws ServletException,IOException{
+  //管理者かどうかのチェック //追記
+    if (checkAdmin()) { //追記
     putRequestScope(AttributeConst.TOKEN,getTokenId());
     putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
 
     forward(ForwardConst.FW_EMP_NEW);
 }
-
+}
 //新規登録を行う
 public void create() throws ServletException,IOException{
-    if(checkToken()) {
+  //管理者かどうかのチェック //追記
+    if (checkAdmin()&& checkToken()) {
         EmployeeView ev = new EmployeeView(
                 null,
                 getRequestParam(AttributeConst.EMP_CODE),
@@ -104,7 +108,8 @@ public void create() throws ServletException,IOException{
 }
 //show
 public void show() throws ServletException, IOException {
-
+  //管理者かどうかのチェック //追記
+    if (checkAdmin()) { //追記
     //idを条件に従業員データを取得する
     EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -119,9 +124,13 @@ public void show() throws ServletException, IOException {
 
     //詳細画面を表示
     forward(ForwardConst.FW_EMP_SHOW);
+    }
 }
 //edit
+
 public void edit() throws ServletException,IOException {
+  //管理者かどうかのチェック //追記
+    if (checkAdmin()) { //追記
     EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
             if(ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
 
@@ -135,12 +144,12 @@ putRequestScope(AttributeConst.TOKEN,getTokenId());
 putRequestScope(AttributeConst.EMPLOYEE,ev);
 
 forward(ForwardConst.FW_EMP_EDIT);
-
+}
 }
 //update
 public void update() throws ServletException,IOException{
 
-    if (checkToken()) {
+    if (checkAdmin()&& checkToken()) {
         EmployeeView ev = new EmployeeView(
                 toNumber(getRequestParam(AttributeConst.EMP_ID)),
                 getRequestParam(AttributeConst.EMP_CODE),
@@ -177,7 +186,7 @@ public void update() throws ServletException,IOException{
 
 //destroy
 public void destroy() throws ServletException, IOException {
-    if(checkToken()) {
+    if (checkAdmin() && checkToken()) { //追記
         service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
         putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
@@ -186,4 +195,19 @@ public void destroy() throws ServletException, IOException {
 
     }
 }
+
+//管理者かどうかチェック
+private boolean checkAdmin() throws ServletException, IOException {
+
+    EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+    if(ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+        forward(ForwardConst.FW_ERR_UNKNOWN);
+        return false;
+
+    } else {
+        return true;
+    }
+    }
+
 }
